# docs/swagger.yaml
definitions:
...
  models.BasicResponse:
    properties:
      message:
        type: string
      result: {}
      status:
        type: boolean
    type: object
...
  models.GenerateLocalRBTRequest:
    properties:
      did:
        type: string
      number_of_tokens:
        type: integer
      start_index:
        type: integer
    type: object
...
  models.RegisterCallBackUrlReq:
    properties:
      callBackURL:
        type: string
      smartContractToken:
        type: string
    type: object
...
  models.TokenDetail:
    properties:
      status:
        type: integer
      token:
        type: string
    type: object
  models.TokenResponse:
    properties:
      message:
        type: string
      result: {}
      status:
        type: boolean
      token_details:
        items:
          $ref: '#/definitions/models.TokenDetail'
        type: array
    type: object
...
  types.SignRespData:
    properties:
      id:
        type: string
      password:
        type: string
      signature:
        description: signature string should be base64
        type: string
    type: object
info:
  contact: {}
  description: Rubix core API to control & manage the node.
  title: Rubix Core
  version: "1.0"
paths:
...
  /api/generate-local-rbt:
    post:
      consumes:
      - application/json
      description: Generates local (test) RBT tokens for the given DID on a local
        network.
      parameters:
      - description: Local RBT generation request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GenerateLocalRBTRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Generate local RBT
      tags:
      - Token
...
  /api/getalltokens:
    get:
      description: Lists all tokens held by the given DID, optionally filtered by
        token type.
      parameters:
      - description: DID (59-char alphanumeric, prefixed with bafybmi)
        in: query
        name: did
        required: true
        type: string
      - description: Token type filter
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
      summary: Get all tokens
      tags:
      - Token
...
  /api/register-callback-url:
    post:
      consumes:
      - application/json
      description: This API will register call back url for when updated come for
        smart contract token
      operationId: register-callback-url
      parameters:
      - description: Register call back URL
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterCallBackUrlReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Get Smart Contract Token Chain Data
      tags:
      - Smart Contract
...
  /rubix/v1/dids/{did}/balances/rbt:
    get:
      consumes:
      - application/json
      description: Retrieves the RBT token balance for a given DID.
      parameters:
      - description: DID (e.g. did:bafybmih3l2emb4s7wbsgakwv4voaqngdirpg5f3kqlheqqsgdg7jthuwaq)
        in: path
        name: did
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Get RBT Balance
      tags:
      - DID
...
  /rubix/v1/signature:
    post:
      consumes:
      - application/json
      description: This API is used to supply the password for the node along with
        the ID generated when Initiate RBT transfer is called.
      operationId: signature-response
      parameters:
      - description: Send input for requested signature
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignRespData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Signature Response
      tags:
      - Signature
...
securityDefinitions:
  SessionToken:
    in: header
    name: Session-Token
    type: apiKey
swagger: "2.0"