# server/tokens.go
package server

import (
	"net/http"
	"regexp"
	"strings"

	"github.com/rubixchain/rubixgoplatform/types"
	"github.com/rubixchain/rubixgoplatform/types/models"
	"github.com/rubixchain/rubixgoplatform/wrapper/ensweb"
)

// APIGetAllTokens godoc
// @Summary      Get all tokens
// @Description  Lists all tokens held by the given DID, optionally filtered by token type.
// @Tags         Token
// @Produce      json
// @Param        did   query     string  true   "DID (59-char alphanumeric, prefixed with bafybmi)"
// @Param        type  query     string  false  "Token type filter"
// @Success      200   {object}  models.TokenResponse
// @Router       /api/getalltokens [get]
func (s *Server) APIGetAllTokens(req *ensweb.Request) *ensweb.Result {
	tokenType := s.GetQuery(req, "type")
	did := s.GetQuery(req, "did")
	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(did)
	if !strings.HasPrefix(did, "bafybmi") || len(did) != 59 || !is_alphanumeric {
		s.log.Error("Invalid DID")
		return s.BasicResponse(req, false, "Invalid DID", nil)
	}
	tr, err := s.c.GetAllTokens(did, tokenType)
	if err != nil {
		return s.BasicResponse(req, false, "Failed to get tokens", nil)
	}
	return s.RenderJSON(req, tr, http.StatusOK)
}

// APIGenerateLocalRBT godoc
// @Summary      Generate local RBT
// @Description  Generates local (test) RBT tokens for the given DID on a local network.
// @Tags         Token
// @Accept       json
// @Produce      json
// @Param        input  body      models.GenerateLocalRBTRequest  true  "Local RBT generation request"
// @Success      200    {object}  models.BasicResponse
// @Router       /api/generate-local-rbt [post]
func (s *Server) APIGenerateLocalRBT(req *ensweb.Request) *ensweb.Result {
	var tr models.GenerateLocalRBTRequest
	err := s.ParseJSON(req, &tr)
	if err != nil {
		return s.BasicResponse(req, false, "Invalid input", nil)
	}
	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(tr.DID)
	if !strings.HasPrefix(tr.DID, "bafybmi") || len(tr.DID) != 59 || !is_alphanumeric {
		s.log.Error("Invalid DID")
		return s.BasicResponse(req, false, "Invalid DID", nil)
	}
	if tr.NumberOfTokens <= 0 {
		s.log.Error("Invalid RBT amount, tokens generated should be a whole number and greater than 0")
		return s.BasicResponse(req, false, "Invalid RBT amount, tokens generated should be a whole number and greater than 0", nil)
	}

	if !s.validateDIDAccess(req, tr.DID) {
		return s.BasicResponse(req, false, "DID does not have an access", nil)
	}
	s.c.AddWebReq(req)
	go s.c.GenerateLocalRBT(req.ID, tr.NumberOfTokens, tr.DID, tr.StartIndex)
	return s.didResponse(req, req.ID)
}

// GetRBTBalance godoc
// @Summary      Get RBT Balance
// @Description  Retrieves the RBT token balance for a given DID.
// @Tags         DID
// @Accept       json
// @Produce      json
// @Param        did  path      string  true  "DID (e.g. did:bafybmih3l2emb4s7wbsgakwv4voaqngdirpg5f3kqlheqqsgdg7jthuwaq)"
// @Success      200  {object}  models.BasicResponse
// @Failure      400  {object}  models.BasicResponse
// @Router       /rubix/v1/dids/{did}/balances/rbt [get]
func (s *Server) APIGetRbtByDid(req *ensweb.Request) *ensweb.Result {
	did := s.GetRouteVar(req, "did")
	if !s.validateDIDAccess(req, did) {
		return s.BasicResponse(req, false, "DID does not have an access", nil)
	}

	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(did)
	if !strings.HasPrefix(did, "bafybmi") || len(did) != 59 || !is_alphanumeric {
		s.log.Error("Invalid DID:", did)
		return s.BasicResponse(req, false, "Invalid DID", nil)
	}
	info, err := s.c.GetRbtByDid(did)
	if err != nil {
		return s.BasicResponse(req, false, err.Error(), nil)
	}
	ac := models.BasicResponse{
		Status:  true,
		Message: "Got account info successfully",
		Result:  info,
	}
	return s.RenderJSON(req, ac, http.StatusOK)
}

// ShowAccount godoc
// @Summary     Signature Response
// @Description This API is used to supply the password for the node along with the ID generated when Initiate RBT transfer is called.
// @Tags        Signature
// @ID 			signature-response
// @Accept      json
// @Produce     json
// @Param 		input body types.SignRespData true "Send input for requested signature"
// @Success      200      {object}  models.BasicResponse
// @Failure      400      {object}  models.BasicResponse
// @Router /rubix/v1/signature [post]
func (s *Server) APISignatureResponse(req *ensweb.Request) *ensweb.Result {
	var resp types.SignRespData
	err := s.ParseJSON(req, &resp)
	if err != nil {
		return s.BasicResponse(req, false, "Invalid input", nil)
	}
	dc := s.c.GetWebReq(resp.ID)
	if dc == nil {
		return s.BasicResponse(req, false, "Invalid request ID", nil)
	}
	s.c.UpateWebReq(resp.ID, req)
	dc.InChan <- resp
	return s.didResponse(req, resp.ID)
}

func (s *Server) APIGenerateFaucetTestToken(req *ensweb.Request) *ensweb.Result {
	var tr models.FaucetRBTGenerateRequest
	err := s.ParseJSON(req, &tr)
	if err != nil {
		return s.BasicResponse(req, false, "Invalid input", nil)
	}
	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(tr.DID)
	if !strings.HasPrefix(tr.DID, "bafybmi") || len(tr.DID) != 59 || !is_alphanumeric {
		s.log.Error("Invalid DID")
		return s.BasicResponse(req, false, "Invalid DID", nil)
	}
	if tr.TokenCount <= 0 {
		s.log.Error("Invalid level number, level should be greater than 0")
		return s.BasicResponse(req, false, "Invalid level number, level should be greater than 0", nil)
	}

	if !s.validateDIDAccess(req, tr.DID) {
		return s.BasicResponse(req, false, "DID does not have an access", nil)
	}
	s.c.AddWebReq(req)
	go s.c.GenerateFaucetTestTokens(req.ID, tr.TokenCount, tr.DID)
	return s.didResponse(req, req.ID)
}